+++ fs/open_options_test.go
